\name{qqplotGenotype}
\alias{qqplotGenotype}
\title{Method "qqplotGenotype"}
\description{
    Creates a Q-Q plot of residuals for each genotype
}
\usage{
    qqplotGenotype(phenList, phenTestResult)
}
\arguments{
    \item{phenList}{instance of the \code{\linkS4class{PhenList}} class}
    \item{phenTestResult}{instance of the \code{\linkS4class{PhenTestResult}} class}
}

\references{
    Placeholder for references
}
\author{Natalja Kurbatova, Natasha Karp, Jeremy Mason}
\seealso{\code{\linkS4class{PhenList}} and \code{\linkS4class{PhenTestResult}}}
\examples{
    test <- PhenList(dataset=read.csv("./PhenStat/extdata/MBAU_data.csv"))
    result <- testDataset(test,depVariable="Bone.Area")
    result <- buildFinalModel(test,result)
    qqplotGenotype(test,result)
}